# app/api-docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Payment Validator API - Documentación del Sistema",
    "description": "API de conciliación de pagos y manejo de Webhooks para la pasarela de procesamiento Payment Validator.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de Desarrollo Local"
    }
  ],
  "paths": {
...
    "/api/transacciones/escanear": {
      "post": {
        "summary": "Escanear y extraer datos de comprobantes (OCR)",
        "description": "Procesa una imagen en base64 usando Google Gemini Vision AI para extraer el monto, la moneda y el ID de Binance Pay del comprobante.",
        "tags": ["Transacciones"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["imagenBase64"],
                "properties": {
                  "imagenBase64": {
                    "type": "string",
                    "description": "Cadena Base64 de la imagen del comprobante (puede incluir el prefijo data:image/png;base64,).",
                    "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Datos extraídos de la imagen con éxito.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Imagen procesada con éxito"
                    },
                    "datos": {
                      "type": "object",
                      "properties": {
                        "monto": { "type": "number", "example": 25.5 },
                        "moneda": { "type": "string", "example": "USDT" },
                        "binance_id_sufijo": { "type": "string", "example": "4310701056" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Falta la imagen en la petición."
          },
          "500": {
            "description": "Error al procesar la imagen con Gemini."
          },
          "502": {
            "description": "El proveedor de IA no respondió (modelo retirado, cuota agotada o fallo de red): reintentar o llenar los datos a mano."
          }
        }
      }
    },
...
  }
}